+++ src/fundamentals/markdownItStatic.tsx
@@ -8,7 +8,7 @@
 
 import markdownit from 'markdown-it'
 import type { JSX, Setter } from 'solid-js'
-import { hljsMarkdownIt } from '../hljsMarkdownIt'
+import { initMarkdownIt } from '../markdownIt'
 import type { Options as MarkdownItOptions } from 'markdown-it'
 
 
@@ -16,17 +16,19 @@ import type { Options as MarkdownItOptions } from 'markdown-it'
  * - On refresh, markdown is rendered on `BE`
  *     - For best SEO please use `<MarkdownItStatic />`
  *     - For dynamic/immediate FE updates use `<MarkdownItDynamic />`
+ * - 🚨 IF not highlighting code THEN `registerHljs` AND `hljsMarkdownItOptions` are not necessary
  * @example
 ```ts
 import mdAppInfo from '@src/md/mdAppInfo.md?raw'
 import { registerHljs } from '@src/init/registerHljs'
 import { MarkdownItStatic } from '@ace/markdownItDynamic'
+import { hljsMarkdownItOptions } from '@ace/hljsMarkdownItOptions'
 
-<MarkdownItStatic content={mdAppInfo} registerHljs={registerHljs} />
+<MarkdownItStatic content={mdAppInfo} registerHljs={registerHljs} options={{ highlight: hljsMarkdownItOptions }} />
 ```
  * @param content - Content to render from markdown to html, can also pass content later by updating the passed in content prop or `md()?.render()`
  * @param setMD - in parent `const [md, setMD] = createSignal<MarkdownIt>()` and then pass `setMD`
- * @param options - Optional, defaults to `defaultMarkdownOptions`, can override one prop at a time b/c we merge
+ * @param options - Optional, requested options will be merged w/ the `defaultMarkdownOptions`
  * @param $div - Optional, props passed to inner wrapper div
  * @param registerHljs - Optional, required when we want code highlighting, registers highlight languages
  */
@@ -35,31 +37,18 @@ export const MarkdownItStatic = (props: {
   content: string,
   /** in parent `const [md, setMD] = createSignal<MarkdownIt>()` and then pass `setMD` */
   setMD?: Setter<markdownit | undefined>
-  /** Optional, defaults to `defaultMarkdownOptions`, can override one prop at a time b/c we merge */
+  /** Optional, requested options will be merged w/ the `defaultMarkdownOptions` */
   options?: MarkdownItOptions
   /** Optional, props passed to inner wrapper div */
   $div?: JSX.HTMLAttributes<HTMLDivElement>,
-  /** Optional, required when we want code highlighting, registers highlight languages */
+  /** Optional, to enable code highlighting pass a function here that registers highlight languages */
   registerHljs?: () => void
 }) => {
-  if (props.registerHljs) props.registerHljs()
-
-  const md = markdownit({
-    ...defaultMarkdownOptions,
-    ...props.options,
-    highlight: props.registerHljs ? hljsMarkdownIt : null
-  })
-
-  props.setMD?.(md)
+  const md = initMarkdownIt({ options: props.options, registerHljs: props.registerHljs, setMD: props.setMD })
 
   return <div {...props.$div} innerHTML={md.render(props.content)} />
 }
 
-export const defaultMarkdownOptions: MarkdownItOptions = {
-  html: true,
-  linkify: true,
-  typographer: true
-}
 
 
 export type MarkdownItStaticProps = Parameters<typeof MarkdownItStatic>[0]